+++ packages/gatsby-starter-hummingbird/gatsby-config.js
@@ -27,7 +27,6 @@ module.exports = {
               && typographyConfig.options.headerFontFamily.join(', '),
           }),
           ...getDesign(),
-          cornerMenuIconColor: '#ffff00',
         },
         typography: {
           ...typographyConfig,
@@ -67,6 +66,16 @@ module.exports = {
           ],
           ...getFooterConfig(),
         },
+      },
+    },
+  ],
+  plugins: [
+    {
+      resolve: 'gatsby-plugin-sentry',
+      options: {
+        dsn: process.env.SENTRY_DSN,
+        environment: process.env.NODE_ENV,
+        enabled: (() => ['production', 'stage'].indexOf(process.env.NODE_ENV) !== -1)(),
       },
     },
   ],