+++ packages/react/src/cards/NodesCard.js
@@ -1,5 +1,4 @@
 import React, { useState, useEffect, useRef } from 'react';
-import styled from '../lib/styled';
 import {
   wide,
   ComplexCard,
@@ -7,6 +6,7 @@ import {
   Loading as _Loading,
   PaginationControls as _PaginationControls,
 } from '@wingscms/crane';
+import styled from '../lib/styled';
 import useQueryString from '../hooks/queryString';
 import { useWings } from '../ctx/Wings';
 import createCard from '../createCard';
@@ -177,7 +177,14 @@ const NodesCardView = ({ text, ...props }) => {
       )
       .catch(err => console.error(err));
 
-    setNodes(res.nodes.edges.map(node => node.node));
+    const itemIds = items.map(item => item.nodeId);
+    const _nodes = res.nodes.edges.map(node => node.node);
+
+    setNodes(
+      type === 'archive'
+        ? _nodes
+        : _nodes.sort((a, b) => itemIds.indexOf(a.id) - itemIds.indexOf(b.id)),
+    );
     setPageInfo(res.nodes.pageInfo);
     setHasLoaded(true);
     setLoading(false);
@@ -195,7 +202,7 @@ const NodesCardView = ({ text, ...props }) => {
           <Loading intent="primary" />
         ) : (
           <FlexGrid
-            divisions={4}
+            divisions={3}
             margins={10}
             alignItems="stretch"
             style={{